Move AI insights to /insights page, remove old placeholder content
- Replace old hardcoded insights page with the AI content analysis UI
- AI patterns, recommendations, and "Go deeper" buttons now live at /insights
- Remove /insights/suggestions subroute (was placeholder data)
- Simplify insights layout (remove subnav)
- Remove AI Insights tab from /analytics (it now lives at /insights)
- Analytics page keeps: Posts, Growth, Audience, Best Times tabs
